% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_helpers.R
\name{legend_location}
\alias{legendBL}
\alias{legendBR}
\alias{legendTL}
\alias{legendTR}
\alias{legend_location}
\title{Put the legend inside a plot}
\usage{
legendTR()

legendBR()

legendBL()

legendTL()
}
\description{
These functions will put the legend in a corner of the main plot and remove the decorations
which set the legend off from the background. They help to reduce the overall size of a plot.

\code{legendTR()}: Legend at top-right

\code{legendBR()}: Legend at bottom-right

\code{legendBL()}: Legend at bottom-left

\code{legendTL)}: Legend at top-left
}
\examples{
ggplot2::qplot(disp, mpg, data=mtcars, color=factor(cyl)) + legendTR()
ggplot2::qplot(disp, mpg, data=mtcars, color=factor(cyl)) + legendBR()
ggplot2::qplot(disp, mpg, data=mtcars, color=factor(cyl)) + legendBL()
ggplot2::qplot(disp, mpg, data=mtcars, color=factor(cyl)) + legendTL()
}

