% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{small_format}
\alias{small_format}
\title{An rmarkdown format based on \code{html_vignette} with custom CSS.
This creates an output file much smaller than \code{html_document}.}
\usage{
small_format(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{html_vignette}}}
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
An rmarkdown format based on \code{html_vignette} with custom CSS.
This creates an output file much smaller than \code{html_document}.
}

