% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_helpers.R
\name{theme_reports}
\alias{theme_reports}
\title{A ggplot2 theme that I like.}
\usage{
theme_reports()
}
\description{
Lighter than the standard gray theme.
}
\examples{
ggplot2::qplot(mtcars$mpg) + theme_reports()
}

